# src/openapi.yaml
openapi: 3.0.0
info:
  title: Hedera Faucet API
  version: 1.2.0
  description: API for dispensing faucet drips for partners
servers:
- url: https://faucet.api.hashport.network
tags:
- name: Faucet
  description: Claim Faucet Drips
- name: Transactions
  description: View partner-specific transaction history
paths:
  /api/drip:
    post:
      tags:
        - Faucet
      summary: Dispenses a drip of HBAR to the specified Hedera Account ID
      description: |
        - Requires a valid API key
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipient_account_id]
              properties:
                recipient_account_id:
                  type: string
                  description: Hedera account ID in dotted format
                  example: "0.0.123456"
      responses:
        '200':
          description: Drip successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimSuccess'
              example:
                code: "DRIP_SUCCESSFUL"
                transactionId: "0.0.1000@1755056331.366704000"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              examples:
                RECIPIENT_ACCOUNT_REQUIRED:
                  summary: Missing recipient_account_id
                  value:
                    code: RECIPIENT_ACCOUNT_REQUIRED
                INVALID_RECIPIENT_ACCOUNT_ID:
                  summary: recipient_account_id is not a valid Hedera account ID
                  value:
                    code: INVALID_RECIPIENT_ACCOUNT_ID
        '401':
          description: Authentication failed
          content:
            application/json:
              examples:
                INVALID_API_KEY:
                  description: The given API Key is invalid.
                  value:
                    code: INVALID_API_KEY
                KEY_REVOKED:
                  description: The given API Key has been revoked, and is no longer usable.
                  value:
                    code: KEY_REVOKED
                KEY_EXPIRED: 
                  description: The given API Key has expired, and is no longer usable.
                  value:
                    code: KEY_EXPIRED
        '403':
          description: The API service has been paused by partner or a system administrator. 
          content:
            application/json:
              example: 
                code: 'SERVICE_PAUSED'

        '402':
          description: The partner's treasury balance is too low to drip. Refill required.
          content:
            application/json:
              example: 
                code: 'INSUFFICIENT_FUNDS'

        '502':
          description: The drip transaction failed. 
          content:
            application/json:
              example: 
                code: 'DRIP_FAILED'
                status: 'FAILED'
                transactionId: '0.0.1000@1755056331.366704000'

        '500':
          description: An internal error has occured. Please contact system administrator.
          content:
            application/json:
              example: 
                code: 'DRIP_INTERNAL_ERROR'
  /api/transactions/:
    get:
      tags:
      - Transactions
      summary: Retrieve all faucet transactions for the authenticated partner
      description: |
        - Requires a valid API key
      security:
        - ApiKeyAuth: []
      responses:
        '401':
          description: Authentication failed
          content:
            application/json:
              examples:
                INVALID_API_KEY:
                  description: The given API Key is invalid.
                  value:
                    code: INVALID_API_KEY
                KEY_REVOKED:
                  description: The given API Key has been revoked, and is no longer usable.
                  value:
                    code: KEY_REVOKED
                KEY_EXPIRED: 
                  description: The given API Key has expired, and is no longer usable.
                  value:
                    code: KEY_EXPIRED
        '403':
          description: The API service has been paused by partner or a system administrator. 
          content:
            application/json:
              example: 
                code: 'SERVICE_PAUSED'
        '200':
          description: List of faucet transaction history entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transactionId:
                      type: string
                      description: Transaction ID (Hashport Txn ID or custom)
                    status:
                      type: string
                      description: Transaction status (SUCCESS, FAILED, etc.)
                    sender_account_id:
                      type: string
                      description: Treasury or sender account ID
                    recipient_account_id:
                      type: string
                      description: Recipient Hedera account ID
                    amount:
                      type: string
                      description: Amount of HBAR dripped
                    timestamp:
                      type: string
                      format: date-time
                      description: Time of the faucet transaction
              example:
                - transactionId: "0.0.5005@1755056331.366704000"
                  status: "SUCCESS"
                  sender_account_id: "0.0.10232596"
                  recipient_account_id: "0.0.9316572"
                  amount: "2.31841213 ℏ"
                  timestamp: "2026-01-17T01:17:14.921Z"
        '500':
          description: Server error when retrieving transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
              example:
                code: "TRANSACTION_INTERNAL_ERROR"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Provide your partner's API Key within the specified headers
      
  schemas:
    ClaimSuccess:
      type: object
      properties:
        code:
          type: string
          enum: [DRIP_SUCCESSFUL]
        transactionId:
          type: string
          description: >
            HashScan **transactionId** string in the form `0.0.ACCOUNT@seconds.nanos`.
            Build full URL as:
            `https://hashscan.io/mainnet/transaction/txId/{hashScan}`
      required: [code, transactionId]